% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ky_dist_geo.R
\docType{data}
\name{ky_hex}
\alias{ky_hex}
\title{Kentucky School District Hex Geography, 2016-2017}
\format{A dataframe with 183 rows and 2 variables:
\describe{
  \item{dist_name}{District name}
  \item{geometry}{sfcMULTIPOLYGON of district boundary information}
}}
\usage{
ky_hex
}
\description{
A dataset with student Kentucky school district geometry data
}
\keyword{datasets}
